+++ src/Order/Order.php
@@ -61,6 +61,10 @@ final class Order
     private $carriagePaid;
     /** @var null|bool */
     private $refunded;
+
+    /** @var float */
+    private $balance;
+
     /**
      * @internal
      *
@@ -98,6 +102,7 @@ final class Order
         $this->isPaid = \array_key_exists('is_paid', $data) ? (bool) $data['is_paid'] : null;
         $this->carriagePaid = $data['carriagePaid'] ?? false;
         $this->refunded = \array_key_exists('refunded', $data) ? (bool) $data['refunded'] : null;
+        $this->balance = (\array_key_exists('balance', $data) === true) ? $data['balance'] : 0;
     }
 
     /**
@@ -268,5 +273,11 @@ final class Order
     public function isRefunded(): ?bool
     {
         return $this->refunded;
+    }
+
+    /** @return float */
+    public function getBalance(): float
+    {
+        return $this->balance;
     }
 }